requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
yfinance>=0.2.35
openpyxl>=3.1.2
